+++ src/TrivyOperator.Dashboard/Domain/TrivyOld/ClusterComplianceReport/Summary.cs
@@ -5,8 +5,8 @@ namespace TrivyOperator.Dashboard.Domain.TrivyOld.ClusterComplianceReport;
 public class Summary
 {
     [JsonPropertyName("failCount")]
-    public long FailCount { get; init; } = 0;
+    public int FailCount { get; init; } = 0;
 
     [JsonPropertyName("passCount")]
-    public long PassCount { get; init; } = 0;
+    public int PassCount { get; init; } = 0;
 }
